services/users/requirements.txt
Flask==0.12.2
Flask-SQLAlchemy==2.3.2
psycopg2==2.7.3.2
Flask-Testing==0.6.2
gunicorn
flask-cors==3.0.3
pyjwt==1.5.3
flask-migrate==2.1.1
flask-bcrypt==0.7.1
pymongo==3.4.0
Flask-PyMongo
Flask-MongoEngine
mongoengine
Flask-MongoAlchemy
redis
rq
Flask-Mail
celery
boto3

